requirements.txt
aiohttp==3.6.2
altgraph==0.17
appdirs==1.4.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==19.3.0
backcall==0.2.0
black==20.8b1
bleach==3.1.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.3
djangorestframework==3.12.2
entrypoints==0.3
filelock==3.0.12
future==0.18.2
idna==2.9
ipykernel==5.3.1
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.1
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
multidict==4.7.6
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.7
notify2==0.3.1
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pathspec==0.7.0
pefile==2019.4.18
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
Pygments==2.6.1
PyInstaller==3.6
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
regex==2020.2.20
requests==2.23.0
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.4.1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.8
virtualenv==20.0.29
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.1
yarl==1.4.2
